Initial commit, outlining employee & shift model
The base repo for a Clojure app as generated by leiningen, along with
the current state of the project outline in rosterbot.org.

The outline has been updated to include the basics of constraint
programming, as well as the basic structure of the model. I'm planning
to iterate over the specified date range and generate employee-day
variables, assigning the constraints as each is generated.